Transferred responsibility for movement from PlayerController -> MovementController
