+++ Assets/MovementController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 34f693fa0d06e104889f6081b3e10304
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
