+++ Assets/Prefabs/Characters/Player.prefab
@@ -663,6 +663,7 @@ GameObject:
   - component: {fileID: 2998625688478654432}
   - component: {fileID: 2998625688478654437}
   - component: {fileID: 2998625688478654436}
+  - component: {fileID: 5395218959478522881}
   - component: {fileID: 539325392}
   - component: {fileID: 2126534817}
   - component: {fileID: 2126534818}
@@ -740,17 +741,29 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: eba6f3e7740cfc64d8318f91e597a077, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &5395218959478522881
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2998625688478654463}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 34f693fa0d06e104889f6081b3e10304, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   baseMovementSpeed: 5
   baseTurnSpeed: 2
-  canMoveRotationThreshold: 1
+  canMoveRotationThreshold: 0.1
   consideredMovementThreshold: 0.1
-  dashForce: 10
+  movementSpeed: 0
+  turnSpeed: 0
+  dashForce: 0
   specialAwakenessDecrease: 0.5
   specialTrailRenderers:
   - {fileID: 1392470110}
   - {fileID: 1530555049}
-  movementSpeed: 3
-  turnSpeed: 2
 --- !u!114 &539325392
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -768,9 +781,6 @@ MonoBehaviour:
   throwPosition: {fileID: 9141267454401288687}
   throwableDestroyTime: 5
   throwMultiplier: 0.08
-  angleIncreaseSpeed: 22.5
-  minThrowAngle: 0
-  maxThrowAngle: 45
   minThrowVelocity: 8
   maxThrowVelocity: 12
 --- !u!114 &2126534817
@@ -801,6 +811,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   charTypeSpeed: 0.05
   textMesh: {fileID: 4161434416140384690}
+  usedForTextCutscenes: 0
 --- !u!1 &3418821350365183018
 GameObject:
   m_ObjectHideFlags: 0
@@ -1360,8 +1371,6 @@ GameObject:
   - component: {fileID: 6213561825049778206}
   - component: {fileID: 2063364861}
   - component: {fileID: 2063364862}
-  - component: {fileID: 4339275362522602481}
-  - component: {fileID: 6020091685612507305}
   - component: {fileID: 784940851590820014}
   - component: {fileID: 6098033672677451275}
   m_Layer: 0
